test(service): add comprehensive unit tests for service layer implementations
- Add CustomerServiceImplTest with order count and customer orders scenarios
- Add OrderServiceImplTest with order total retrieval and exception handling
- Add OrderMessagingServiceImplTest with message processing and error scenarios
- Implement proper mock configuration using @MockBean and @InjectMocks
- Cover edge cases: null parameters, not found exceptions, and runtime errors
- Use verifyNoInteractions() for proper mock verification in exception scenarios
- Align test data with actual domain models (Order constructor, BigDecimal usage)
- Test pagination scenarios for customer orders endpoint
- Validate proper exception propagation from use cases to service layer
- Update monetary values from Double to BigDecimal for precision compliance

Test coverage includes:
- Success scenarios for all service operations
- Null parameter handling and validation
- Exception propagation from domain layer
- Proper mapper interaction verification
- Edge cases and error conditions

Benefits:
- Complete unit test coverage for service layer implementations
- Proper validation of service behavior and exception handling
- Clean test structure following testing best practices
- Consistent mock usage and verification patterns
- Type safety with BigDecimal for monetary calculations

+++ src/main/java/com/btg/challenge/orders/app/mapper/OrderMapper.java
@@ -26,7 +26,7 @@ public class OrderMapper {
         OrderSummary summary = new OrderSummary();
         summary.setOrderId(order.getOrderId());
         summary.setCustomerId(order.getCustomerId());
-        summary.setTotalAmount(order.getTotalAmount().doubleValue());
+        summary.setTotalAmount(order.getTotalAmount());
         summary.setItemsCount(order.getItemsCount());
         summary.setCreatedAt(order.getCreatedAt());
